ui 1차 통일 및 정산/problem-ocr 3차 수정

+++ backend/src/main/java/com/ieum/backend/domain/settlement/dto/response/SettlementSummaryResponse.java
@@ -2,8 +2,8 @@ package com.ieum.backend.domain.settlement.dto.response;
 
 public record SettlementSummaryResponse(
         Long tutorId,
-        Integer totalAmount,       // 누적 총 정산 금액
-        Integer transferredAmount, // 송금 완료된 총 금액
-        Integer pendingAmount,     // 송금 대기 + 계산 완료 금액
-        Integer settlementCount    // 정산 건수
+        long totalAmount,       // 누적 총 정산 금액(원) — 합계는 long으로(오버플로 방지)
+        long transferredAmount, // 송금 완료된 총 금액
+        long pendingAmount,     // 송금 대기 + 계산 완료 금액
+        long settlementCount    // 정산 건수
 ) {}
